Use uintptr for file descriptor to avoid so many casts

+++ userfaultfd.go
@@ -11,8 +11,8 @@ import (
 	"golang.org/x/sys/unix"
 )
 
-func ioctl(fd int, op uintptr, arg unsafe.Pointer) error {
-	_, _, errno := unix.Syscall(unix.SYS_IOCTL, uintptr(fd), op, uintptr(arg))
+func ioctl(fd uintptr, op uintptr, arg unsafe.Pointer) error {
+	_, _, errno := unix.Syscall(unix.SYS_IOCTL, fd, op, uintptr(arg))
 	if errno != 0 {
 		return os.NewSyscallError("ioctl", errno)
 	}
@@ -50,7 +50,7 @@ func Open(flags int) (*os.File, error) {
 
 // ApiHandshake negotiates the userfaultfd API version and features.
 // Returns the negotiated info or an error.
-func ApiHandshake(fd int, features uint64) (*UffdioApi, error) {
+func ApiHandshake(fd uintptr, features uint64) (*UffdioApi, error) {
 	api := &UffdioApi{Api: UFFD_API, Features: features}
 	if err := ioctl(fd, UFFDIO_API, unsafe.Pointer(api)); err != nil {
 		return nil, err
@@ -59,7 +59,7 @@ func ApiHandshake(fd int, features uint64) (*UffdioApi, error) {
 }
 
 // Continue resolves a minor page fault for the given range.
-func Continue(fd int, start uintptr, length int, mode int) error {
+func Continue(fd uintptr, start uintptr, length int, mode int) error {
 	if !HaveIoctlContinue {
 		return ErrMissingIoctl
 	}
@@ -72,7 +72,7 @@ func Continue(fd int, start uintptr, length int, mode int) error {
 
 // Copy resolves a page fault by copying content from src to dst.
 // Returns the number of bytes copied or an error.
-func Copy(fd int, dst, src uintptr, length int, mode int) (int64, error) {
+func Copy(fd uintptr, dst, src uintptr, length int, mode int) (int64, error) {
 	c := &UffdioCopy{Dst: uint64(dst), Src: uint64(src), Len: uint64(length), Mode: uint64(mode)}
 	if err := ioctl(fd, UFFDIO_COPY, unsafe.Pointer(c)); err != nil {
 		return 0, err
@@ -82,7 +82,7 @@ func Copy(fd int, dst, src uintptr, length int, mode int) (int64, error) {
 
 // Move moves pages from src to dst within the same process.
 // Returns the number of bytes/pages moved or an error.
-func Move(fd int, dst, src uintptr, length int, mode int) (int64, error) {
+func Move(fd uintptr, dst, src uintptr, length int, mode int) (int64, error) {
 	if !HaveIoctlMove {
 		return 0, ErrMissingIoctl
 	}
@@ -96,7 +96,7 @@ func Move(fd int, dst, src uintptr, length int, mode int) (int64, error) {
 // Poison marks pages in the given range as poisoned. Subsequent accesses
 // may generate SIGBUS or other behaviour depending on kernel semantics.
 // Returns the number of pages/bytes updated (as reported by the kernel).
-func Poison(fd int, start uintptr, length int, mode int) (int64, error) {
+func Poison(fd uintptr, start uintptr, length int, mode int) (int64, error) {
 	if !HaveIoctlPoison {
 		return 0, ErrMissingIoctl
 	}
@@ -109,7 +109,7 @@ func Poison(fd int, start uintptr, length int, mode int) (int64, error) {
 
 // Register registers a memory range for userfaultfd handling with the specified mode.
 // Returns the registration info or an error.
-func Register(fd int, start uintptr, length int, mode int) (*UffdioRegister, error) {
+func Register(fd uintptr, start uintptr, length int, mode int) (*UffdioRegister, error) {
 	reg := &UffdioRegister{Range: UffdioRange{Start: uint64(start), Len: uint64(length)}, Mode: uint64(mode)}
 	if err := ioctl(fd, UFFDIO_REGISTER, unsafe.Pointer(reg)); err != nil {
 		return nil, err
@@ -118,7 +118,7 @@ func Register(fd int, start uintptr, length int, mode int) (*UffdioRegister, err
 }
 
 // Unregister unregisters a previously registered range.
-func Unregister(fd int, start uintptr, length int) error {
+func Unregister(fd uintptr, start uintptr, length int) error {
 	r := &UffdioRange{Start: uint64(start), Len: uint64(length)}
 	if err := ioctl(fd, UFFDIO_UNREGISTER, unsafe.Pointer(r)); err != nil {
 		return err
@@ -127,7 +127,7 @@ func Unregister(fd int, start uintptr, length int) error {
 }
 
 // Wake wakes up blocked page faults in the given range.
-func Wake(fd int, start uintptr, length int) error {
+func Wake(fd uintptr, start uintptr, length int) error {
 	r := &UffdioRange{Start: uint64(start), Len: uint64(length)}
 	if err := ioctl(fd, UFFDIO_WAKE, unsafe.Pointer(r)); err != nil {
 		return err
@@ -136,7 +136,7 @@ func Wake(fd int, start uintptr, length int) error {
 }
 
 // WriteProtect enables or disables write protection on a range.
-func WriteProtect(fd int, start uintptr, length int, mode int) error {
+func WriteProtect(fd uintptr, start uintptr, length int, mode int) error {
 	if !HaveIoctlWriteProtect {
 		return ErrMissingIoctl
 	}
@@ -149,7 +149,7 @@ func WriteProtect(fd int, start uintptr, length int, mode int) error {
 
 // Zeropage resolves a page fault by zero-filling the memory range.
 // Returns the length zeroed or an error.
-func Zeropage(fd int, start uintptr, length int, mode int) (int64, error) {
+func Zeropage(fd uintptr, start uintptr, length int, mode int) (int64, error) {
 	z := &UffdioZeropage{Range: UffdioRange{Start: uint64(start), Len: uint64(length)}, Mode: uint64(mode)}
 	if err := ioctl(fd, UFFDIO_ZEROPAGE, unsafe.Pointer(z)); err != nil {
 		return 0, err
